# parser/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dmarc-parser API",
    "version": "2.0.0",
    "description": "REST API of the DMARC aggregate report parser & ingestion service.\n\n### Authentication\nEvery `/api/v1/*` route (except `/api/v1/openapi.json`) requires an API\nkey from `PARSER_API_KEYS`, sent either as `Authorization: Bearer <key>`\nor `X-API-Key: <key>`. Keys carry scopes from `{read, ingest, admin}`;\neach operation documents its required scope via `x-required-scope`.\nMissing/invalid key \u2192 401; key without the required scope \u2192 403.\n\n### Rate limiting\nPer-key token bucket (`PARSER_RATE_LIMIT_RPS`, default 10 req/s,\nburst 3x). Exceeding it returns 429 with a `Retry-After` header.\n\n### Conventions\nAll timestamps are UTC, RFC 3339. Date-only values accepted by `since` /\n`until` parameters use `YYYY-MM-DD`. Errors are JSON\n`{\"error\": \"message\"}`.\n"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8081",
      "description": "Host loopback"
    },
    {
      "url": "http://dmarcparser:8080",
      "description": "From containers on the mxsentinel_default / dmarcparser_default networks"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/ingest": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Ingest DMARC aggregate report(s)",
        "description": "Body may be raw XML, gzip, or zip (any Content-Type), or multipart/form-data with a file field. Zip archives may contain multiple reports; all are processed. Duplicates (same domain + report_id) are no-ops returning the existing serial.\n",
        "operationId": "ingestReport",
        "x-required-scope": "ingest",
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "At least one new report stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "200": {
            "description": "Every report was a duplicate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/poll": {
      "post": {
        "tags": [
          "poller"
        ],
        "summary": "Trigger an immediate IMAP poll cycle (async)",
        "operationId": "triggerPoll",
        "x-required-scope": "admin",
        "responses": {
          "202": {
            "description": "Poll triggered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "const": "poll triggered"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/status": {
      "get": {
        "tags": [
          "poller"
        ],
        "summary": "Poller state, lifetime counters and DB totals",
        "operationId": "getStatus",
        "x-required-scope": "read",
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/reports": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List stored reports",
        "operationId": "listReports",
        "x-required-scope": "read",
        "parameters": [
          {
            "$ref": "#/components/parameters/domain"
          },
          {
            "$ref": "#/components/parameters/org"
          },
          {
            "$ref": "#/components/parameters/since"
          },
          {
            "$ref": "#/components/parameters/until"
          },
          {
            "$ref": "#/components/parameters/limit500"
          },
          {
            "$ref": "#/components/parameters/offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Report summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reports": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ReportSummary"
                      }
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/reports/{serial}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "One report including all records",
        "operationId": "getReport",
        "x-required-scope": "read",
        "parameters": [
          {
            "$ref": "#/components/parameters/serial"
          }
        ],
        "responses": {
          "200": {
            "description": "Report detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key from PARSER_API_KEYS as a bearer token. Key scopes: read, ingest, admin (see each operation's x-required-scope)."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Same API key via the X-API-Key header."
      }
    },
    "parameters": {
      "domain": {
        "name": "domain",
        "in": "query",
        "description": "Filter by policy-published domain",
        "schema": {
          "type": "string"
        }
      },
      "org": {
        "name": "org",
        "in": "query",
        "description": "Filter by reporting organisation",
        "schema": {
          "type": "string"
        }
      },
      "since": {
        "name": "since",
        "in": "query",
        "description": "Lower time bound (RFC 3339 or YYYY-MM-DD)",
        "schema": {
          "type": "string"
        }
      },
      "until": {
        "name": "until",
        "in": "query",
        "description": "Upper time bound (RFC 3339 or YYYY-MM-DD)",
        "schema": {
          "type": "string"
        }
      },
      "limit500": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 500,
          "default": 50
        }
      },
...
      "offset": {
        "name": "offset",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
      "serial": {
        "name": "serial",
        "in": "path",
        "required": true,
        "description": "Report serial",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "API key lacks the required scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Per-key rate limit exceeded (see Retry-After)",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying",
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "IngestResponse": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "serial": {
                  "type": "integer",
                  "format": "int64"
                },
                "domain": {
                  "type": "string"
                },
                "org": {
                  "type": "string"
                },
                "report_id": {
                  "type": "string"
                },
                "records": {
                  "type": "integer"
                },
                "messages": {
                  "type": "integer",
                  "format": "int64"
                },
                "duplicate": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      },
      "StatusResponse": {
        "type": "object",
        "properties": {
          "poller": {
            "type": "object",
            "properties": {
              "enabled": {
                "type": "boolean"
              },
              "interval": {
                "type": "string"
              },
              "last_poll": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "last_success": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "last_error": {
                "type": "string"
              }
            }
          },
          "counters": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int64"
            }
          },
          "database": {
            "type": "object",
            "properties": {
              "reports": {
                "type": "integer",
                "format": "int64"
              },
              "records": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        }
      },
      "ReportSummary": {
        "type": "object",
        "properties": {
          "serial": {
            "type": "integer",
            "format": "int64"
          },
          "org": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "report_id": {
            "type": "string"
          },
          "date_begin": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "date_end": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "records": {
            "type": "integer",
            "format": "int64"
          },
          "messages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Record": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "source_ip": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "disposition": {
            "type": [
              "string",
              "null"
            ]
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "dkim_domain": {
            "type": [
              "string",
              "null"
            ]
          },
          "dkim_result": {
            "type": [
              "string",
              "null"
            ]
          },
          "spf_domain": {
            "type": [
              "string",
              "null"
            ]
          },
          "spf_result": {
            "type": [
              "string",
              "null"
            ]
          },
          "spf_align": {
            "type": "string"
          },
          "dkim_align": {
            "type": "string"
          },
          "header_from": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ReportDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ReportSummary"
          },
          {
            "type": "object",
            "properties": {
              "email": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "extra_contact_info": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "policy_adkim": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "policy_aspf": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "policy_p": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "policy_sp": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "policy_pct": {
                "type": [
                  "integer",
                  "null"
                ]
              },
              "record_rows": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Record"
                }
              }
            }
          }
        ]
      },
...
    }
  }
}